hv: Add check to ensure vcpu_make_request and signal_event are consistent
Currently the vcpu_make_request and signal_event in vcpu_lock_instr_emulation
does not check whether target VCPU is up and running. This can cause
problems because when VCPU is created but not launched,
vcpu_make_request will not trigger wait_event on target VCPU, but
signal_event may still execute to reduce the counter.

This patch adds a check before vcpu_make_request and signal_event to
make sure the request and signal are issued after target VCPU is up.

Tracked-On: #6502

+++ hypervisor/include/common/event.h
@@ -4,7 +4,7 @@
 
 struct sched_event {
 	spinlock_t lock;
-	int8_t nqueued;
+	int32_t nqueued;
 	struct thread_object* waiting_thread;
 };
